refactor: 패키지 계층 정리 & 게임 진행 SRP 강화
- Pakage 계층 정리(분리)
  - car
  - factory
  - game
    - generator
  - view

- 기존의 RacingGame 클래스를 GameEngine으로 명칭 변경
- GameEngine의 playTurn() 메서드가 난수 생성/전진 여부 결정 두 가지
  책임을 지님
- 그래서, 난수 생성 책임을 NumberGenerator 인터페이스로 분리 후
  RandomNumberGenerator 클래스로 구현
- GameEngine은 generator를 생성자로 주입 받음
