+++ src/main/java/racingcar/domain/factory/CarRegistrar.java
@@ -1,4 +1,6 @@
-package racingcar.domain;
+package racingcar.domain.factory;
+
+import racingcar.domain.car.RacingCar;
 
 public class CarRegistrar {
     private static final int MAX_NAME_LENGTH = 5;
