+++ src/main/java/racingcar/domain/game/generator/NumberGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.domain.game.generator;
+
+public interface NumberGenerator {
+    int generate();
+}
